Moved image section to app detail, refactored code
