+++ src/apps/console/routes/_main+/$account+/$project+/$environment+/app+/$app+/settings+/environment/route.tsx
@@ -4,9 +4,7 @@ import {
   createAppEnvPage,
   useAppState,
 } from '~/console/page-components/app-states';
-import Wrapper from '~/console/components/wrapper';
-import { useUnsavedChanges } from '~/root/lib/client/hooks/use-unsaved-changes';
-import { Button } from '~/components/atoms/button';
+import AppWrapper from '~/console/page-components/app/app-wrapper';
 import { EnvironmentVariables } from '../../../../new-app/app-environment-variables';
 import { ConfigMounts } from '../../../../new-app/app-config-mount';
 
@@ -18,7 +16,6 @@ export interface IAppDialogValue {
 
 const SettingEnvironment = () => {
   const { envPage, setEnvPage } = useAppState();
-  const { setPerformAction, hasChanges, loading } = useUnsavedChanges();
   const items: {
     label: string;
     value: createAppEnvPage;
@@ -34,48 +31,22 @@ const SettingEnvironment = () => {
   ];
 
   return (
-    <div>
-      <Wrapper
-        secondaryHeader={{
-          title: 'Environment',
-          action: hasChanges && (
-            <div className="flex flex-row items-center gap-lg">
-              <Button
-                disabled={loading}
-                variant="basic"
-                content="Discard changes"
-                onClick={() => setPerformAction('discard-changes')}
-              />
-              <Button
-                disabled={loading}
-                content={loading ? 'Committing changes.' : 'View changes'}
-                loading={loading}
-                onClick={() => setPerformAction('view-changes')}
-              />
-            </div>
-          ),
-        }}
-      >
-        <ExtendedFilledTab
-          value={envPage}
-          onChange={setEnvPage}
-          items={items}
-        />
-        <AnimatePresence mode="wait">
-          <motion.div
-            key={envPage || 'empty'}
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            transition={{ duration: 0.15 }}
-            className="flex flex-col gap-6xl w-full"
-          >
-            {envPage === 'environment_variables' && <EnvironmentVariables />}
-            {envPage === 'config_mounts' && <ConfigMounts />}
-          </motion.div>
-        </AnimatePresence>
-      </Wrapper>
-    </div>
+    <AppWrapper title="Environment">
+      <ExtendedFilledTab value={envPage} onChange={setEnvPage} items={items} />
+      <AnimatePresence mode="wait">
+        <motion.div
+          key={envPage || 'empty'}
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          exit={{ opacity: 0 }}
+          transition={{ duration: 0.15 }}
+          className="flex flex-col gap-6xl w-full"
+        >
+          {envPage === 'environment_variables' && <EnvironmentVariables />}
+          {envPage === 'config_mounts' && <ConfigMounts />}
+        </motion.div>
+      </AnimatePresence>
+    </AppWrapper>
   );
 };
