+++ src/apps/console/routes/_main+/$account+/$project+/$environment+/app+/$app+/settings+/general/route.tsx
@@ -3,28 +3,40 @@ import { useEffect } from 'react';
 import { TextInput } from '~/components/atoms/input';
 import { Box } from '~/console/components/common-console-components';
 import { useAppState } from '~/console/page-components/app-states';
-import { keyconstants } from '~/console/server/r-utils/key-constants';
-import useForm from '~/root/lib/client/hooks/use-form';
+import useForm, { dummyEvent } from '~/root/lib/client/hooks/use-form';
 import Yup from '~/root/lib/server/helpers/yup';
 import { parseName } from '~/console/server/r-utils/common';
-import Wrapper from '~/console/components/wrapper';
-import { Button } from '~/components/atoms/button';
-import { useUnsavedChanges } from '~/root/lib/client/hooks/use-unsaved-changes';
+import AppWrapper from '~/console/page-components/app/app-wrapper';
+import RepoSelector from '~/console/page-components/app/components';
+import { registryHost } from '~/root/lib/configs/base-url.cjs';
+import { useParams } from '@remix-run/react';
 
 const SettingGeneral = () => {
-  const { app, setApp } = useAppState();
-  const { setPerformAction, hasChanges, loading } = useUnsavedChanges();
+  const { app, setApp, activeContIndex } = useAppState();
 
-  const { values, errors, handleChange, submit, resetValues } = useForm({
+  const { account } = useParams();
+
+  const { values, errors, handleChange, submit, setValues } = useForm({
     initialValues: {
       name: parseName(app),
       displayName: app.displayName,
-      description: app.metadata?.annotations?.[keyconstants.description] || '',
+      imageUrl: app.spec.containers[activeContIndex]?.image || '',
+      manualRepo: '',
     },
     validationSchema: Yup.object({
       name: Yup.string().required(),
       displayName: Yup.string().required(),
-      description: Yup.string(),
+      imageUrl: Yup.string(),
+      manualRepo: Yup.string().when(['imageUrl'], ([imageUrl], schema, c) => {
+        const regex = /^(\w+):(\w+)$/;
+        if (!imageUrl) {
+          if (!c.value) {
+            return schema.required('Image is required');
+          }
+          return schema.matches(regex, 'Invalid image format');
+        }
+        return schema;
+      }),
     }),
 
     onSubmit: async (val) => {
@@ -36,10 +48,19 @@ const SettingGeneral = () => {
             name: val.name,
             annotations: {
               ...(a.metadata?.annotations || {}),
-              [keyconstants.description]: val.description,
             },
           },
           displayName: val.displayName,
+          spec: {
+            ...a.spec,
+            containers: [
+              {
+                ...(a.spec.containers?.[0] || {}),
+                image: val.imageUrl || val.manualRepo,
+                name: 'container-0',
+              },
+            ],
+          },
         };
       });
     },
@@ -49,65 +70,58 @@ const SettingGeneral = () => {
     submit();
   }, [values]);
 
-  useEffect(() => {
-    if (!hasChanges) {
-      resetValues();
-    }
-  }, [hasChanges]);
-
   return (
-    <div>
-      <Wrapper
-        secondaryHeader={{
-          title: 'General',
-          action: hasChanges && (
-            <div className="flex flex-row items-center gap-lg">
-              <Button
-                disabled={loading}
-                variant="basic"
-                content="Discard changes"
-                onClick={() => setPerformAction('discard-changes')}
-              />
-              <Button
-                disabled={loading}
-                content={loading ? 'Committing changes.' : 'View changes'}
-                loading={loading}
-                onClick={() => setPerformAction('view-changes')}
-              />
-            </div>
-          ),
-        }}
-      >
-        <Box title="Application detail">
-          <div className="flex flex-row items-center gap-3xl">
-            <div className="flex-1">
-              <TextInput
-                label="Application name"
-                error={!!errors.displayName}
-                message={errors.displayName}
-                onChange={handleChange('displayName')}
-                value={values.displayName}
-              />
-            </div>
-            <div className="flex-1">
-              <TextInput
-                label="Application ID"
-                value={values.name}
-                suffixIcon={<CopySimple />}
-                disabled
-              />
-            </div>
+    <AppWrapper title="General">
+      <Box title="Application detail">
+        <div className="flex flex-row items-center gap-3xl">
+          <div className="flex-1">
+            <TextInput
+              label="Application name"
+              error={!!errors.displayName}
+              message={errors.displayName}
+              onChange={handleChange('displayName')}
+              value={values.displayName}
+            />
           </div>
-          <TextInput
-            label="Description"
-            error={!!errors.description}
-            message={errors.description}
-            value={values.description}
-            onChange={handleChange('description')}
-          />
-        </Box>
-      </Wrapper>
-    </div>
+          <div className="flex-1">
+            <TextInput
+              label="Application ID"
+              value={values.name}
+              suffixIcon={<CopySimple />}
+              disabled
+            />
+          </div>
+        </div>
+        <RepoSelector
+          tag={values.imageUrl.split(':')[1]}
+          repo={
+            values.imageUrl
+              .replace(`${registryHost}/${account}/`, '')
+              .split(':')[0]
+          }
+          onClear={() => {
+            setValues((v) => {
+              return {
+                ...v,
+                imageUrl: '',
+                manualRepo: '',
+              };
+            });
+          }}
+          textValue={values.manualRepo}
+          onTextChanged={(e) => {
+            handleChange('manualRepo')(e);
+            handleChange('imageUrl')(dummyEvent(''));
+          }}
+          onValueChange={({ repo, tag }) => {
+            handleChange('imageUrl')(
+              dummyEvent(`${registryHost}/${account}/${repo}:${tag}`)
+            );
+          }}
+          error={errors.manualRepo}
+        />
+      </Box>
+    </AppWrapper>
   );
 };
 export default SettingGeneral;
