+++ src/apps/console/routes/_main+/$account+/$project+/$environment+/app+/$app+/settings+/scaling/route.tsx
@@ -3,16 +3,13 @@ import Slider from '~/components/atoms/slider';
 import { useAppState } from '~/console/page-components/app-states';
 import useForm, { dummyEvent } from '~/root/lib/client/hooks/use-form';
 import Yup from '~/root/lib/server/helpers/yup';
-import Wrapper from '~/console/components/wrapper';
-import { useUnsavedChanges } from '~/root/lib/client/hooks/use-unsaved-changes';
-import { Button } from '~/components/atoms/button';
 import { Checkbox } from '~/components/atoms/checkbox';
+import AppWrapper from '~/console/page-components/app/app-wrapper';
 
 const SettingCompute = () => {
   const { app, setApp } = useAppState();
-  const { setPerformAction, hasChanges, loading } = useUnsavedChanges();
 
-  const { values, handleChange, submit, resetValues } = useForm({
+  const { values, handleChange, submit } = useForm({
     initialValues: {
       autoscaling: app.spec.hpa?.enabled || false,
       minReplicas: app.spec.hpa?.minReplicas || 1,
@@ -70,133 +67,105 @@ const SettingCompute = () => {
     submit();
   }, [values]);
 
-  useEffect(() => {
-    if (!hasChanges) {
-      resetValues();
-    }
-  }, [hasChanges]);
-
   return (
-    <div>
-      <Wrapper
-        secondaryHeader={{
-          title: 'Scaling',
-          action: hasChanges && (
-            <div className="flex flex-row items-center gap-lg">
-              <Button
-                disabled={loading}
-                variant="basic"
-                content="Discard changes"
-                onClick={() => setPerformAction('discard-changes')}
-              />
-              <Button
-                disabled={loading}
-                content={loading ? 'Committing changes.' : 'View changes'}
-                loading={loading}
-                onClick={() => setPerformAction('view-changes')}
-              />
-            </div>
-          ),
-        }}
-      >
-        <div className="flex flex-col gap-3xl">
-          <Checkbox
-            label="Autoscaling"
-            checked={values.autoscaling}
-            onChange={(val) => {
-              handleChange('autoscaling')(dummyEvent(val));
-            }}
-          />
-          {values.autoscaling ? (
-            <div className="flex flex-col gap-3xl">
-              <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default">
-                <div className="flex flex-row gap-lg items-center">
-                  <div className="bodyMd-medium text-text-default">
-                    Select min and max replicas
-                  </div>
-                  <code className="bodyMd text-text-soft flex-1 text-end">
-                    {values.minReplicas || 75} min - {values.maxReplicas || 75}{' '}
-                    max
-                  </code>
-                </div>
-                <Slider
-                  step={1}
-                  min={1}
-                  max={10}
-                  value={[values.minReplicas, values.maxReplicas]}
-                  onChange={(value) => {
-                    if (Array.isArray(value)) {
-                      handleChange('minReplicas')(dummyEvent(value[0]));
-                      handleChange('maxReplicas')(dummyEvent(value[1]));
-                    }
-                  }}
-                />
-              </div>
-              <div className="flex flex-row justify-between gap-3xl">
-                <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default w-full">
-                  <div className="flex flex-row gap-lg items-center">
-                    <div className="bodyMd-medium text-text-default">
-                      Select CPU
-                    </div>
-                    <code className="bodyMd text-text-soft flex-1 text-end">
-                      {values.cpuThreshold || 75}% CPU
-                    </code>
-                  </div>
-                  <Slider
-                    step={1}
-                    min={50}
-                    max={95}
-                    value={values.cpuThreshold}
-                    onChange={(value) => {
-                      handleChange('cpuThreshold')(dummyEvent(value));
-                    }}
-                  />
-                </div>
-                <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default w-full">
-                  <div className="flex flex-row gap-lg items-center">
-                    <div className="bodyMd-medium text-text-default">
-                      Select Memory
-                    </div>
-                    <code className="bodyMd text-text-soft flex-1 text-end">
-                      {values.memoryThreshold || 75}% Memory
-                    </code>
-                  </div>
-                  <Slider
-                    step={1}
-                    min={50}
-                    max={95}
-                    value={values.memoryThreshold}
-                    onChange={(value) => {
-                      handleChange('memoryThreshold')(dummyEvent(value));
-                    }}
-                  />
-                </div>
-              </div>
-            </div>
-          ) : (
+    <AppWrapper title="Scaling">
+      <div className="flex flex-col gap-3xl">
+        <Checkbox
+          label="Autoscaling"
+          checked={values.autoscaling}
+          onChange={(val) => {
+            handleChange('autoscaling')(dummyEvent(val));
+          }}
+        />
+        {values.autoscaling ? (
+          <div className="flex flex-col gap-3xl">
             <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default">
               <div className="flex flex-row gap-lg items-center">
                 <div className="bodyMd-medium text-text-default">
-                  Select replicas
+                  Select min and max replicas
                 </div>
                 <code className="bodyMd text-text-soft flex-1 text-end">
-                  {values.replicas || 1} replicas
+                  {values.minReplicas || 75} min - {values.maxReplicas || 75}{' '}
+                  max
                 </code>
               </div>
               <Slider
                 step={1}
                 min={1}
                 max={10}
-                value={values.replicas}
+                value={[values.minReplicas, values.maxReplicas]}
                 onChange={(value) => {
-                  handleChange('replicas')(dummyEvent(value));
+                  if (Array.isArray(value)) {
+                    handleChange('minReplicas')(dummyEvent(value[0]));
+                    handleChange('maxReplicas')(dummyEvent(value[1]));
+                  }
                 }}
               />
             </div>
-          )}
-        </div>
-      </Wrapper>
-    </div>
+            <div className="flex flex-row justify-between gap-3xl">
+              <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default w-full">
+                <div className="flex flex-row gap-lg items-center">
+                  <div className="bodyMd-medium text-text-default">
+                    Select CPU
+                  </div>
+                  <code className="bodyMd text-text-soft flex-1 text-end">
+                    {values.cpuThreshold || 75}% CPU
+                  </code>
+                </div>
+                <Slider
+                  step={1}
+                  min={50}
+                  max={95}
+                  value={values.cpuThreshold}
+                  onChange={(value) => {
+                    handleChange('cpuThreshold')(dummyEvent(value));
+                  }}
+                />
+              </div>
+              <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default w-full">
+                <div className="flex flex-row gap-lg items-center">
+                  <div className="bodyMd-medium text-text-default">
+                    Select Memory
+                  </div>
+                  <code className="bodyMd text-text-soft flex-1 text-end">
+                    {values.memoryThreshold || 75}% Memory
+                  </code>
+                </div>
+                <Slider
+                  step={1}
+                  min={50}
+                  max={95}
+                  value={values.memoryThreshold}
+                  onChange={(value) => {
+                    handleChange('memoryThreshold')(dummyEvent(value));
+                  }}
+                />
+              </div>
+            </div>
+          </div>
+        ) : (
+          <div className="flex flex-col gap-md p-2xl rounded border border-border-default bg-surface-basic-default">
+            <div className="flex flex-row gap-lg items-center">
+              <div className="bodyMd-medium text-text-default">
+                Select replicas
+              </div>
+              <code className="bodyMd text-text-soft flex-1 text-end">
+                {values.replicas || 1} replicas
+              </code>
+            </div>
+            <Slider
+              step={1}
+              min={1}
+              max={10}
+              value={values.replicas}
+              onChange={(value) => {
+                handleChange('replicas')(dummyEvent(value));
+              }}
+            />
+          </div>
+        )}
+      </div>
+    </AppWrapper>
   );
 };
 export default SettingCompute;
