+++ src/apps/console/routes/_main+/$account+/$project+/$environment+/new-app/app-config-mount.tsx
@@ -11,7 +11,6 @@ import { useAppState } from '~/console/page-components/app-states';
 import useForm from '~/root/lib/client/hooks/use-form';
 import Yup from '~/root/lib/server/helpers/yup';
 import { NonNullableString } from '~/root/lib/types/common';
-import { useUnsavedChanges } from '~/root/lib/client/hooks/use-unsaved-changes';
 import {
   ArrowRight,
   ChevronLeft,
@@ -193,7 +192,6 @@ export const ConfigMounts = () => {
   const [showCSDialog, setShowCSDialog] = useState<IShowDialog>(null);
 
   // for updating
-  const { hasChanges } = useUnsavedChanges();
 
   const entry = Yup.object({
     type: Yup.string().oneOf(['config', 'secret']).notRequired(),
@@ -208,12 +206,7 @@ export const ConfigMounts = () => {
     mountPath: Yup.string().required(),
   });
 
-  const {
-    values,
-    setValues,
-    submit,
-    resetValues: resetAppValue,
-  } = useForm({
+  const { values, setValues, submit } = useForm({
     initialValues: getContainer().volumes,
     validationSchema: Yup.array(entry),
     onSubmit: (val) => {
@@ -228,13 +221,6 @@ export const ConfigMounts = () => {
   useEffect(() => {
     submit();
   }, [values]);
-
-  // for updating
-  useEffect(() => {
-    if (!hasChanges) {
-      resetAppValue();
-    }
-  }, [hasChanges]);
 
   const addEntry = (val: IConfigMount) => {
     console.log('here', val);
@@ -294,7 +280,6 @@ export const ConfigMounts = () => {
         }),
     }),
     onSubmit: () => {
-      console.log('helre', eValues);
       if (eValues.value) {
         const ev: IConfigMount = {
           refName: eValues.value.refName,
