+++ src/apps/console/routes/_main+/$account+/$project+/$environment+/new-app/app-review.tsx
@@ -115,14 +115,14 @@ const AppReview = () => {
           </div>
         </ReviewComponent>
 
-        {app.spec.nodeSelector[keyconstants.nodepoolName] !== '' && (
+        {!!app.spec.nodeSelector?.[keyconstants.nodepoolName] && (
           <ReviewComponent title="Nodepool Details" onEdit={() => {}}>
             <div className="flex flex-col p-xl gap-md rounded border border-border-default">
               <div className="bodyMd-semibold text-text-default">
                 Nodepool Selector
               </div>
               <div className="bodySm text-text-soft">
-                {app.spec.nodeSelector[keyconstants.nodepoolName]}
+                {app.spec.nodeSelector?.[keyconstants.nodepoolName]}
               </div>
             </div>
           </ReviewComponent>
